Refactor Agent's initialization methods.
